Refactor: moved all the functions to a separate file
made a new file for each function in productUtils folder
