+++ src/utils/productUtils/filterByRange.jsx
@@ -1,0 +1,7 @@
+function filterByRange(products, rangeLimit) {
+  return products.filter(product =>
+    product['price'] <= rangeLimit ? product : false
+  )
+}
+
+export { filterByRange }
