+++ src/utils/productUtils/index.jsx
@@ -1,0 +1,11 @@
+import { sortByPrice } from './sortByPrice'
+import { filterProductsByRating } from './filterProductsByRating'
+import { filterProductsByCategory } from './filterProductsByCategory'
+import { filterByRange } from './filterByRange'
+
+export {
+  sortByPrice,
+  filterProductsByCategory,
+  filterByRange,
+  filterProductsByRating,
+}
