+++ src/utils/productUtils/sortByRating.jsx
@@ -1,0 +1,7 @@
+function sortRatingInDescending(products) {
+  return [...products].sort(
+    (firstItem, secondItem) => secondItem.rating - firstItem.rating
+  )
+}
+
+export { sortRatingInDescending }
